% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_facts.R
\docType{data}
\name{cat_facts}
\alias{cat_facts}
\title{Get a random cat fact about Custurd}
\format{
An object of class \code{character} of length 7.
}
\usage{
cat_facts
}
\value{
A character string containing a random cat fact about Custard.
}
\description{
This function randomly selects and returns a cat fact about Custard from the
list of cat facts provided in the package.
}
\examples{
get_cat_fact()

}
\keyword{datasets}
