% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cat_name.R
\name{get_cat_name}
\alias{get_cat_name}
\title{Get the name of My cat, Custard}
\usage{
get_cat_name()
}
\value{
A character string containing the name "Custard".
}
\description{
This function returns the name of My cat, Custard.
}
