% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cat_name.R
\name{get_cat_name}
\alias{get_cat_name}
\title{Get the name of your cat, Custurd.}
\usage{
get_cat_name()
}
\value{
A character string containing the name "Custurd".
}
\description{
This function returns the name of your cat, Custurd.
}
\examples{
get_cat_name()

}
